requirements.txt
crewai

langchain
langchain-anthropic
langchain-community
langchain-groq
langchain-openai

beautifulsoup4
feedparser
python-dateutil
pytube
pytz

Jinja2
pandas
python-docx
python-dotenv
requests